+++ src/Gaffer/BackgroundTask.cpp
@@ -90,9 +90,9 @@ const ScriptNode *scriptNode( const GraphComponent *subject )
 		}
 	}
 
-	// The `GafferUI::View` classes house internal nodes which might not be
-	// directly connected to the graph. This hack recovers the ScriptNode from
-	// the node the view is currently connected to.
+	// The `GafferUI::View` and `GafferUI.Editor` classes house internal nodes
+	// which might not be directly connected to the graph. This hack recovers
+	// the ScriptNode from such classes.
 	while( subject )
 	{
 		if( subject->isInstanceOf( "GafferUI::View" ) )
@@ -100,6 +100,13 @@ const ScriptNode *scriptNode( const GraphComponent *subject )
 			if( auto inPlug = subject->getChild<Plug>( "in" ) )
 			{
 				return scriptNode( inPlug->getInput() );
+			}
+		}
+		else if( subject->isInstanceOf( "GafferUI::Editor::Settings" ) )
+		{
+			if( auto scriptPlug = subject->getChild<Plug>( "__scriptNode" ) )
+			{
+				return scriptNode( scriptPlug->getInput() );
 			}
 		}
 		subject = subject->parent();